requirement.txt
Django==2.0
django-vote==2.2.0
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
